Squashed 'externals/nitro/' changes from c8ecbe9ae..547d0aa9f
547d0aa9f NITRO-2.11.4 (#575)
fe309c4ba Merge branch 'main' of github.com:mdaus/nitro
10efa9990 latest from CODA-OSS (#574)
d08f1c0a1 CRLF
db5d3d484 latest from CODA-OSS
98c755048 CRLF
aa1482543 CRLF
bfdbe69a4 Latest from CODA-OSS (#573)
9e4ce0b58 latest from CODA-OSS (#572)
227a8a8f4 trust coda-oss for right -std flags
13869687e latest from CODA-OSS (#571)
5724d8c18 latest from CODA-OSS (#570)
fb794f0fe latest from CODA-OSS (#569)
7a6132ba0 update files changes in cpp17 branch (#568)
de91d4977 Fix bug creating NITFException (#567)
225273436 fix YAML for 'main'
5d1c83d11 single project for unittests (#566)
102a019db latest from CODA-OSS (known broken build) (#565)
b4ae2d429 match YAML from coda-oss
176bcaf6d build NITRO.SLN using msbuild (#562)
85e9043b8 latest from CODA-OSS (#561)
49ec50325 use new "PlatformToolkit" special environment variable (#560)
6c06e3711 latest from CODA-OSS (#559)
cee9feb42 latest from CODA-OSS (#558)
3f01809fa latest from coda-oss (#557)
471fb3fc1 Update test_j2k_loading++.cpp
b12caf2fc latest from CODA-OSS (#556)
0cd432624 use sys::make_span (#555)
05dae18a3 patch to build other projects
8f974e995 NITRO-2.11.3 (#554)
870aa6afd update to coda-oss 2023-06-05 (#553)
2fd7a0bfa latest from coda-oss (#552)
0eecce004 invoke() utility to reduce code duplication (#550)
59fb02fe9 latest from coda-oss (#551)
9fbf2b7b8 Fill out adapter free block which is used for nitf decompression (#549)
089ba0b5b latest from coda-oss
3b52f0025 latest from coda-oss (#547)
90c6263e2 latest from coda-oss (#544)
90d513ac5 latest from coda-oss (#543)

git-subtree-dir: externals/nitro
git-subtree-split: 547d0aa9ffcf34ad94f9ea71aa415f39837b2e68

+++ externals/coda-oss/modules/c++/sys/source/ResourceCheck.cpp
@@ -22,6 +22,7 @@
 
 #include <algorithm>
 #include <sys/ResourceCheck.h>
+#include <gsl/gsl.h>
 
 namespace sys
 {
@@ -40,7 +41,7 @@ size_t getAvailableMem(size_t systemMemBytes,
                        size_t reservedBytes,
                        double margin)
 {
-    const auto relativeAvailable = static_cast<size_t>(systemMemBytes * margin);
+    const auto relativeAvailable = gsl::narrow_cast<size_t>(gsl::narrow_cast<double>(systemMemBytes) * margin);
     if (reservedBytes > systemMemBytes)
     {
         return 0;
